+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Changed the name of the package's main export to `Adapter`. The export is
+  also exported under its previous name (`PreactAdapter`) for backwards
+  compatibility
+
 ## [1.8.0] - 2019-03-12
 
 ### Changed